+++ theme/template/posts/speaker.php
@@ -1,3 +1,5 @@
+<?php get_header(); ?>
+
 <header class="header header--subpage">
   <div class="container">
     <div class="row">
@@ -33,3 +35,5 @@
     </div>
   </div>
 </section>
+
+<?php get_footer(); ?>